+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,48 @@
 # @valbuild/cli
 
+## 0.120.1
+
+### Patch Changes
+
+- Updated dependencies [[`6f318d4`](https://github.com/valbuild/val/commit/6f318d406295b772e721bf463283f47e2822e996)]:
+  - @valbuild/server@0.120.1
+  - @valbuild/language-server@0.120.1
+
+## 0.120.0
+
+### Patch Changes
+
+- Updated dependencies [[`c2d3c0e`](https://github.com/valbuild/val/commit/c2d3c0e6c2010c0a94c725d9dbaa618998773e8a)]:
+  - @valbuild/core@0.120.0
+  - @valbuild/shared@0.120.0
+  - @valbuild/language-server@0.120.0
+  - @valbuild/server@0.120.0
+
+## 0.119.0
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @valbuild/server@0.119.0
+  - @valbuild/language-server@0.119.0
+
+## 0.118.0
+
+### Patch Changes
+
+- Updated dependencies [[`198ba8b`](https://github.com/valbuild/val/commit/198ba8bd8e6c921660e97f5cd26fb17f2d5f3f95)]:
+  - @valbuild/server@0.118.0
+  - @valbuild/shared@0.118.0
+  - @valbuild/language-server@0.118.0
+
+## 0.117.1
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @valbuild/server@0.117.1
+  - @valbuild/language-server@0.117.1
+
 ## 0.117.0
 
 ### Patch Changes